Compartmentalize into smaller components
