+++ src/module/apps/lcp-manager/LCPSelector.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+</script>
+
+<div class="lancer-header lancer-primary major" style="grid-area: 1/1/2/3;">Import/Update LCP</div>
+<input id="lcp-file" type="file" name="lcp-up" class="lcp-up" accept=".lcp" />
+
+<style lang="scss">
+</style>
